+++ lib/kettle/test/rspec.rb
@@ -5,3 +5,7 @@ require_relative "external"
 require "kettle/test"
 
 require_relative "internal"
+
+# A gem's test harness should do require "rake" if it is their dependency,
+#   and they define a rake task they want to test.
+require_relative "support/shared_contexts/with_rake" if defined?(Rake)